+++ resources/views/components/input.blade.php
@@ -1,13 +1,8 @@
 <div class="@if(isset($fieldWrapperClass)){{$fieldWrapperClass}}@endif">
-    <label for="{{ $key }}" class="lfb-label">
-        {{ $label }}
-        @if ((!isset($mode) || (isset($mode) and $mode != 'view')) and isset($rules) and array_key_exists($key, $rules) && str_contains($rules[$key], 'required|'))
-            <sup>*</sup>
-        @endif
-    </label>
+    @include('lara-forms-builder::includes.field-label')
     @if (isset($mode) && ($mode == 'view' || $mode == 'confirm'))
         <div class="lfb-input-wrapper lfb-input-readonly">
-            <input type="@if(isset($inputType)){{$inputType}}@else{{'text'}}@endif" name="{{ $key }}" id="{{ $key }}" x-ref="field"
+            <input type="@if(isset($inputType)){{$inputType}}@else{{'text'}}@endif" name="{{ $key }}" id="{{ $key }}"
             value="@if ($model->$key || is_numeric($model->$key) || $this->$key || is_numeric($this->$key)){{ $model->$key ?? $this->$key }}@else - @endif"
             class="lfb-input lfb-disabled" readonly disabled>
             @if(isset($formWarnings) && array_key_exists($key, $formWarnings))