+++ resources/views/includes/submit-button.blade.php
@@ -1,0 +1,3 @@
+<button wire:click="checkAndSave" type="button" class="{{$this->getPrimaryButtonClasses()}}" @if($disableSaveButton) disabled @endif wire:loading.attr="disabled">
+    {{ (isset($mode) && $mode == 'confirm') ? __('Confirm') : $submitButtonLabel }}
+</button>