+++ resources/views/includes/tabs.blade.php
@@ -1,58 +1,74 @@
-{{-- TODO: add css to config --}}
-<div x-data="{ tab: @entangle('activeTab') }" class="bg-white shadow rounded-md">
-    <div class="divide-y divide-gray-200 lg:grid lg:grid-cols-12 lg:divide-y-0 lg:divide-x">
+<div x-data="{ tab: @entangle('activeTab') }" class="lfb-tabs-wrapper">
+    <div class="lfb-tab-container">
         {{-- Tabs Links --}}
-        <aside class="py-4 lg:col-span-2">
-            <div class="flex flex-shrink-0 items-center px-6 mb-6">
-                <div class="text-3xl font-semibold text-c_gray-800">{{ $formTitle }}</div>
-            </div>
-            <nav x-model="tab">
-                @foreach($fields as $field)
-                    <li x-bind:class="[ tab == '{{ $field['tab']['key'] }}' ? 'bg-c_gray-100 border-c_gray-300' : '']" class="text-c_gray-800 border-transparent hover:bg-c_gray-200/50 hover:border-c_gray-300 border-l-4 px-3 py-2 flex items-center text-sm font-medium cursor-pointer">
-                        <a x-on:click.prevent="tab='{{ $field['tab']['key'] }}'">
-                            {{ $field['tab']['title'] }}
-                        </a>
-                    </li>
-                @endforeach
-            </nav>
-        </aside>
-
-        {{-- Tabs Content --}}
-        <div class="lg:col-span-10 relative" wire:target="save" wire:loading.class="opacity-50">
-            <div class="relative">
-                @error('tabWarning')
-                    <div class="shadow mx-4 mt-6 overflow-hidden rounded-md" role="alert">
-                        <div class="flex">
-                            <div class="bg-cyan-500 w-14 text-center p-2">
-                                <div class="flex justify-center h-full items-center">
-                                    <x-heroicon-o-exclamation-circle  class="h-6 w-6 text-white" />
+        @if($isMultiStep)
+            <aside class="lfb-steps-nav-wrapper">
+                <div class="lfb-steps-nav-container">
+                    <nav class="lfb-steps-nav">
+                        @foreach($fields as $field)
+                            <div class="lfb-step-nav">
+                                <div x-bind:class="[ tab == '{{ $field['key'] }}' ? 'lfb-step-nav-title-active' : '']" class="lfb-step-nav-title">
+                                    {{ $field['title'] }}
                                 </div>
                             </div>
-                            <div class="bg-gray-100 border-r-4 border-cyan-500 w-full p-4">
-                                <div>
-                                <p class="text-gray-600 text-sm">{{ $message }}</p>
-                                </div>
+                        @endforeach
+                    </nav>
+                </div>
+            </aside>
+        @else
+            <aside class="lfb-tab-nav-wrapper">
+                <div class="lfb-tab-nav-container">
+                    @if(!empty($formTitle))
+                        <div class="lfb-tab-nav-title-wrapper">
+                            <div class="lfb-tab-nav-title">{{ $formTitle }}</div>
+                        </div>
+                    @endif
+                    <nav class="lfb-tab-nav-links" x-model="tab">
+                        @foreach($fields as $field)
+                            <div class="lfb-tab-nav-link-item">
+                                <a x-bind:class="[ tab == '{{ $field['key'] }}' ? 'lfb-tab-nav-link-active' : '']" class="lfb-tab-nav-link" x-on:click.prevent="tab='{{ $field['key'] }}'">
+                                    {{ $field['title'] }}
+                                </a>
+                            </div>
+                        @endforeach
+                    </nav>
+                </div>
+            </aside>
+        @endif
+
+        {{-- Tabs Content --}}
+        <div class="lfb-tab-content-wrapper" wire:target="save" wire:loading.class="lfb-tab-content-wrapper-loading-class">
+            <div class="lfb-tab-content-container">
+                @error('tabWarning')
+                    <div class="lfb-tab-error-wrapper" role="alert">
+                        <div class="lfb-tab-error-container">
+                            <div class="lfb-tab-error-icon flex justify-center items-center">
+                                @if(isset($tabWarningIcon))
+                                    {!! $tabWarningIcon !!}
+                                @else
+                                 <svg xmlns="http://www.w3.org/2000/svg" fill="none" viewBox="0 0 24 24" stroke-width="1.5" stroke="currentColor" aria-hidden="true">
+                                     <path stroke-linecap="round" stroke-linejoin="round" d="M12 9v3.75m9-.75a9 9 0 11-18 0 9 9 0 0118 0zm-9 3.75h.008v.008H12v-.008z"></path>
+                                 </svg>
+                                @endif
+                            </div>
+                            <div class="lfb-tab-error-message">
+                                <p>{{ $message }}</p>
                             </div>
                         </div>
                     </div>
                 @enderror
                 @foreach($fields as $fieldKey => $field)
-                    <div x-show="tab == '{{ $field['tab']['key'] }}'" wire:key="{{ 'tab-'. md5($field['tab']['key']) }}" x-cloak class="py-6">
-                        <div class="px-4">
-                            <div>
-                                <h2 class="text-xl font-normal leading-6 text-c_gray-800">{{ $field['tab']['title'] }}</h2>
-                            </div>
+                    <div x-show="tab == '{{ $field['key'] }}'" wire:key="{{ 'tab-'. md5($field['key']) }}" x-cloak class="lfb-tab-content-item">
+                            <h2 class="lfb-tab-content-item-title">{{ $field['title'] }}</h2>
                             @include('lara-forms-builder::includes.fields', [
                                     'fields' => [
-                                        $fieldKey => $field['tab']['content']
+                                        $fieldKey => $field['content']
                                     ]
                                 ]
                             )
-                        </div>
                     </div>
                 @endforeach
             </div>
-
             @include('lara-forms-builder::includes.buttons')
         </div>
     </div>